requirements.txt
# ── Core ──
python-dotenv==1.1.0
pydantic>=2.0

# ── OCR ──
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow>=10.0

# ── RAG ──
langchain==0.3.25
langchain-community==0.3.24
langchain-groq==0.3.2
sentence-transformers==3.4.1
faiss-cpu

# ── API ──
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# ── Frontend ──
streamlit==1.45.1
requests>=2.31

# ── Dev ──
pytest==8.3.5
httpx==0.28.1
